fix(rbac): rework currentUser alias to cover ownershipEntityRefs
