+++ plugins/rbac-backend/package.json
@@ -42,13 +42,13 @@
     "@backstage/plugin-permission-backend": "^0.5.46",
     "@backstage/plugin-permission-common": "^0.8.0",
     "@backstage/plugin-permission-node": "^0.8.0",
-    "@dagrejs/graphlib": "^2.1.13",
     "@janus-idp/backstage-plugin-rbac-common": "1.8.1",
     "@janus-idp/backstage-plugin-rbac-node": "1.4.0",
     "@janus-idp/backstage-plugin-audit-log-node": "1.4.0",
     "casbin": "^5.27.1",
     "chokidar": "^3.6.0",
     "csv-parse": "^5.5.5",
+    "@dagrejs/graphlib": "^2.1.13",
     "express": "^4.18.2",
     "express-promise-router": "^4.1.1",
     "knex": "^3.0.0",
